Address Copilot follow-up review
- Fixes a data race in FoundryModuleStatsFetcher#call: rate_limited used
  to be a local variable mutated from inside Parallel.map worker threads,
  which is non-deterministic depending on thread scheduling. Each thread
  now returns its own {module:, rate_limited:} pair and the top-level
  flag is derived from the results after all threads finish.
- Extracts a shared useFoundryModuleStats hook so the dashboard panel
  and the full stats page share the Auth0-token + fetch + refresh
  semantics. Eliminates the duplicated load() callback that lived in
  both components and was already starting to drift.
- Removes the unused stub_http helper from the fetcher spec and reworks
  the misleading "empty/error rows" comment that suggested an error
  state when the rows are actually plain empty rows.
