+++ app/javascript/bundles/DungeonMasterCampaignManager/pages/admin-dashboard/foundry-module-stats/FoundryModuleStatsPage.tsx
@@ -1,5 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
-import { useAuth0 } from '@auth0/auth0-react';
+import React from 'react';
 import PageContainer from '../../../containers/PageContainer';
 import PageTitle from '../../../components/PageTitle/PageTitle';
 import Frame from '../../../components/Frame/Frame';
@@ -8,7 +7,7 @@ import { Colors } from '../../../utilities/enums';
 import SummaryBar from './components/SummaryBar';
 import InstallDonut from './components/InstallDonut';
 import ModuleAccordion from './components/ModuleAccordion';
-import { FoundryModuleStatsResponse } from './types';
+import useFoundryModuleStats from './useFoundryModuleStats';
 import {
   ChartPlaceholder,
   ErrorMsg,
@@ -19,44 +18,8 @@ import {
   Wrapper,
 } from './FoundryModuleStatsPage.styles';
 
-const ENDPOINT = '/v1/foundry-module-stats.json';
-
 const FoundryModuleStatsPage: React.FC = () => {
-  const { getAccessTokenSilently } = useAuth0();
-  const [data, setData] = useState<FoundryModuleStatsResponse | null>(null);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  const load = useCallback(
-    async (refresh = false) => {
-      setLoading(true);
-      setError(null);
-      try {
-        const token = await getAccessTokenSilently();
-        const url = refresh ? `${ENDPOINT}?refresh=1` : ENDPOINT;
-        const response = await fetch(url, {
-          headers: {
-            Accept: 'application/json',
-            Authorization: `Bearer ${token}`,
-          },
-        });
-        if (!response.ok) {
-          throw new Error(`HTTP ${response.status}`);
-        }
-        const json = (await response.json()) as FoundryModuleStatsResponse;
-        setData(json);
-      } catch (e) {
-        setError(e instanceof Error ? e.message : 'Failed to load stats');
-      } finally {
-        setLoading(false);
-      }
-    },
-    [getAccessTokenSilently],
-  );
-
-  useEffect(() => {
-    void load(false);
-  }, [load]);
+  const { data, loading, error, reload } = useFoundryModuleStats();
 
   const lastUpdated = data?.fetched_at
     ? `Last updated: ${new Date(data.fetched_at).toLocaleString()}`
@@ -75,9 +38,7 @@ const FoundryModuleStatsPage: React.FC = () => {
             <Button
               color={Colors.primary}
               title={loading ? 'Loading…' : 'Refresh'}
-              onClick={() => {
-                void load(true);
-              }}
+              onClick={reload}
               disabled={loading}
               isLoading={loading}
             />
